requirements.txt
streamlit>=1.40.0
python-dotenv==1.0.0
openai==1.3.5
pillow>=8.0.0
